033.apigateway-openapi-cognito-auth/tests/requirements.txt
# テスト用Pythonパッケージ
boto3>=1.26.0
pytest>=7.0.0
python-dotenv>=1.0.0
requests>=2.28.0
